// src/types/cause.ts
// Generated by https://quicktype.io

export interface CausesResponse {
  causes: Cause[];
  meta: Meta;
}

export interface Cause {
  name: string;
  description: string;
  distance: null;
  uuid: string;
  city: null | string;
  state: null | string;
  country: Country;
  zip_code: null | string;
  address: string;
  city_state_zip: string;
  latitude: number | null;
  longitude: number | null;
  featured: boolean;
  slogan: null | string;
  keywords: string;
  links: Link[];
  featured_at: null;
  deleted_at: null;
  created_at: string;
  updated_at: string;
  logo: Logo | null;
  thumbnail: Thumbnail | null;
  categories: string[];
  shortener: null | string;
  details_url: string;
  join_url: string;
}

export enum Country {
  Us = 'US'
}

export interface Link {
  url: string;
  type: string;
}

export interface Logo {
  filename: string;
  content_type: string;
  byte_size: number;
  url: string;
}

export interface Thumbnail {
  url: string;
}

export interface Meta {
  has_more: boolean;
  limit: number;
  offset: number;
  count: number;
}